Removing redundant properties from AndroidManifest.xml and renaming package to fix build problems.
