+++ samples/treasurehunt/src/main/AndroidManifest.xml
@@ -23,7 +23,7 @@
             android:icon="@drawable/ic_launcher"
             android:label="@string/app_name">
         <activity
-                android:name=".TreasureHuntActivity"
+                android:name="com.google.vrtoolkit.cardboard.samples.TreasureHuntActivity"
                 android:label="@string/app_name"
                 android:screenOrientation="landscape"
                 android:configChanges="orientation|keyboardHidden|screenSize" >
