+++ samples/treasurehunt/src/main/java/com/google/vrtoolkit/cardboard/samples/TreasureHuntActivity.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package com.google.vrtoolkit.cardboard.samples.treasurehunt;
+package com.google.vrtoolkit.cardboard.samples;
 
 import com.google.vrtoolkit.cardboard.CardboardActivity;
 import com.google.vrtoolkit.cardboard.CardboardView;
@@ -22,6 +22,7 @@ import com.google.vrtoolkit.cardboard.Eye;
 import com.google.vrtoolkit.cardboard.HeadTransform;
 import com.google.vrtoolkit.cardboard.Viewport;
 import com.google.vrtoolkit.cardboard.audio.CardboardAudioEngine;
+import com.google.vrtoolkit.cardboard.samples.treasurehunt.R;
 
 import android.content.Context;
 import android.opengl.GLES20;
