+++ samples/treasurehunt/src/main/java/com/google/vrtoolkit/cardboard/samples/WorldLayoutData.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package com.google.vrtoolkit.cardboard.samples.treasurehunt;
+package com.google.vrtoolkit.cardboard.samples;
 
 /**
  * Contains vertex, normal and color data.
